@@ -6,24 +6,44 @@ import com.bookapp.core.exception.ValidationException;
 import com.bookapp.core.port.CommentRepositoryPort;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
 
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
 
+/**
+ * Сервіс коментарів із використанням Spring DI та конфігурації
+ */
+@Service
 public class CommentService {
     private static final Logger log = LoggerFactory.getLogger(CommentService.class);
-    private static final int MAX_AUTHOR_LENGTH = 100;
-    private static final int MAX_TEXT_LENGTH = 1000;
-    private static final long DELETE_ALLOWED_HOURS = 24;
+
+    // Ін'єкція значень із application.properties
+    @Value("${comments.max-author-length:100}")
+    private int maxAuthorLength;
+
+    @Value("${comments.max-text-length:1000}")
+    private int maxTextLength;
+
+    @Value("${comments.delete-allowed-hours:24}")
+    private long deleteAllowedHours;
 
     private final CommentRepositoryPort repository;
 
+    /**
+     * Ін'єкція через конструктор (найкращий спосіб)
+     */
+    @Autowired
     public CommentService(CommentRepositoryPort repository) {
         this.repository = repository;
+        log.info("CommentService initialized");
     }
 
     public List<Comment> getCommentsByBookId(Long bookId) {
+        log.debug("Getting comments for book_id={}", bookId);
         return repository.findCommentsByBookId(bookId);
     }
 
@@ -41,8 +61,9 @@ public class CommentService {
                 .orElseThrow(() -> new BusinessException("Comment not found"));
 
         Duration age = Duration.between(comment.getCreatedAt(), LocalDateTime.now());
-        if (age.toHours() > DELETE_ALLOWED_HOURS) {
-            throw new BusinessException("Cannot delete comment older than 24 hours");
+        if (age.toHours() > deleteAllowedHours) {
+            throw new BusinessException(
+                    "Cannot delete comment older than " + deleteAllowedHours + " hours");
         }
 
         boolean deleted = repository.deleteComment(commentId);
@@ -55,14 +76,16 @@ public class CommentService {
         if (author == null || author.trim().isEmpty()) {
             throw new ValidationException("Author is required");
         }
-        if (author.length() > MAX_AUTHOR_LENGTH) {
-            throw new ValidationException("Author too long (max " + MAX_AUTHOR_LENGTH + " chars)");
+        if (author.length() > maxAuthorLength) {
+            throw new ValidationException(
+                    "Author too long (max " + maxAuthorLength + " chars)");
         }
         if (text == null || text.trim().isEmpty()) {
             throw new ValidationException("Text is required");
         }
-        if (text.length() > MAX_TEXT_LENGTH) {
-            throw new ValidationException("Text too long (max " + MAX_TEXT_LENGTH + " chars)");
+        if (text.length() > maxTextLength) {
+            throw new ValidationException(
+                    "Text too long (max " + maxTextLength + " chars)");
         }
     }
 }
